+++ pkg/lib/ratelimit/storage.go
@@ -6,5 +6,5 @@ import (
 )
 
 type Storage interface {
-	Update(ctx context.Context, key string, period time.Duration, burst int, delta int) (ok bool, timeToAct time.Time, err error)
+	Update(ctx context.Context, key string, period time.Duration, burst int, delta float64) (ok bool, timeToAct time.Time, err error)
 }
